fix: preserve SDK protocol edge cases
